backend/requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.14,<2.0

# ── Task Queue ────────────────────────────────────────────────────────────────
celery>=5.4,<6.0
redis>=5.0,<6.0

# ── Auth / JWT ────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
httpx>=0.27,<1.0

# ── File Storage ──────────────────────────────────────────────────────────────
boto3>=1.35,<2.0

# ── Vector DB ─────────────────────────────────────────────────────────────────
qdrant-client>=1.12,<2.0

# ── AI / LLM ─────────────────────────────────────────────────────────────────
langchain>=0.3,<1.0
langchain-community>=0.3,<1.0
ollama>=0.4,<1.0
crewai>=0.80,<1.0

# ── Real-time ─────────────────────────────────────────────────────────────────
websockets>=12.0,<16.0
python-socketio>=5.11,<6.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-multipart>=0.0.20,<1.0
aiofiles>=24.0,<25.0
orjson>=3.10,<4.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0
